+++ nextstep-di/src/test/java/nextstep/di/factory/example/BeanFactoryTest.java
@@ -1,12 +1,12 @@
 package nextstep.di.factory.example;
 
-import nextstep.di.ClassPathBeanScanner;
-import nextstep.di.ConfigurationBeanScanner;
 import nextstep.di.factory.BeanFactory;
 import nextstep.di.factory.example.config.MyJdbcTemplate;
 import nextstep.di.factory.example.controller.QnaController;
 import nextstep.di.factory.example.controller.TestQnaController;
 import nextstep.di.factory.example.service.MyQnaService;
+import nextstep.di.scanner.ClassPathBeanScanner;
+import nextstep.di.scanner.ConfigurationBeanScanner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
